Update lpips.py
When I using the lpips.LPIPS(net='vgg'), I found that the output would be the shape like [[[[0.3333]]]], which lead a bad using experience and the result could not be used in the last loss combination operation. However, add squeeze to delete the "[]" automatically could solve the problem mentioned above.